move ipcMain handlers to separate folder
