+++ app/app.html
@@ -32,7 +32,9 @@
         if (process.env.START_HOT) {
           // Dynamically insert the bundled app script in the renderer process
           const port = process.env.PORT || 1212;
-          scripts.push(`http://localhost:${port}/dist/mainWindow.renderer.dev.js`);
+          scripts.push(
+            `http://localhost:${port}/dist/mainWindow.renderer.dev.js`
+          );
         } else {
           scripts.push('./dist/mainWindow.renderer.prod.js');
         }
