+++ app/locales/de/translation.missing.json
@@ -1,3 +1,1 @@
-{
-
-}
+{}
