+++ app/main.dev.ts
@@ -10,23 +10,17 @@
  */
 import 'core-js/stable';
 import 'regenerator-runtime/runtime';
-import { Display } from 'electron/main';
 import path from 'path';
-import { app, BrowserWindow, ipcMain, screen, shell } from 'electron';
+import { app, BrowserWindow, shell } from 'electron';
 import settings from 'electron-settings';
 import i18n from './configs/i18next.config';
 import signalingServer from './server';
 import MenuBuilder from './menu';
 import initGlobals from './utils/mainProcessHelpers/initGlobals';
-import ConnectedDevicesService from './features/ConnectedDevicesService';
-import RoomIDService from './server/RoomIDService';
-import SharingSession from './features/SharingSessionService/SharingSession';
-import getDeskreenGlobal from './utils/mainProcessHelpers/getDeskreenGlobal';
 import AppUpdater from './utils/AppUpdater';
 import installExtensions from './utils/installExtensions';
 import getNewVersionTag from './utils/getNewVersionTag';
-
-const v4IPGetter = require('internal-ip').v4;
+import initIpcMainHandlers from './main/ipcMainHandlers';
 
 export default class DeskreenApp {
   mainWindow: BrowserWindow | null = null;
@@ -102,74 +96,7 @@ export default class DeskreenApp {
   }
 
   initIpcMain() {
-    ipcMain.on('client-changed-language', async (_, newLangCode) => {
-      i18n.changeLanguage(newLangCode);
-      await settings.set('appLanguage', newLangCode);
-    });
-
-    ipcMain.handle('get-signaling-server-port', () => {
-      if (this.mainWindow === null) return;
-      this.mainWindow.webContents.send(
-        'sending-port-from-main',
-        signalingServer.port
-      );
-    });
-
-    ipcMain.handle('get-all-displays', () => {
-      return screen.getAllDisplays();
-    });
-
-    ipcMain.handle('get-display-size-by-display-id', (_, displayID: string) => {
-      const display = screen.getAllDisplays().find((d: Display) => {
-        return `${d.id}` === displayID;
-      });
-
-      if (display) {
-        return display.size;
-      }
-      return undefined;
-    });
-
-    ipcMain.handle('main-window-onbeforeunload', () => {
-      const deskreenGlobal = getDeskreenGlobal();
-      deskreenGlobal.connectedDevicesService = new ConnectedDevicesService();
-      deskreenGlobal.roomIDService = new RoomIDService();
-      deskreenGlobal.sharingSessionService.sharingSessions.forEach(
-        (sharingSession: SharingSession) => {
-          sharingSession.denyConnectionForPartner();
-          sharingSession.destroy();
-        }
-      );
-
-      deskreenGlobal.rendererWebrtcHelpersService.helpers.forEach(
-        (helperWindow) => {
-          helperWindow.close();
-        }
-      );
-
-      deskreenGlobal.sharingSessionService.waitingForConnectionSharingSession = null;
-      deskreenGlobal.rendererWebrtcHelpersService.helpers.clear();
-      deskreenGlobal.sharingSessionService.sharingSessions.clear();
-    });
-
-    ipcMain.handle('get-latest-version', () => {
-      return this.latestVersion;
-    });
-
-    ipcMain.handle('get-current-version', () => {
-      return this.appVersion;
-    });
-
-    ipcMain.handle('get-local-lan-ip', async () => {
-      if (
-        process.env.RUN_MODE === 'dev' ||
-        process.env.NODE_ENV === 'production'
-      ) {
-        const ip = await v4IPGetter();
-        return ip;
-      }
-      return '255.255.255.255';
-    });
+    initIpcMainHandlers(this.mainWindow, this.latestVersion, this.appVersion);
   }
 
   async createWindow() {
